🚚 (consent-das): Rename service dir to services
