+++ apps/consent-das/src/routers/participants.ts
@@ -19,8 +19,8 @@
 
 import { Router } from 'express';
 
-import { getParticipant, getParticipants } from '../service/search.js';
-import { createParticipant } from '../service/create.js';
+import { getParticipant, getParticipants } from '../services/search.js';
+import { createParticipant } from '../services/create.js';
 import logger from '../logger.js';
 
 // TODO: update JSDoc comments when custom error handling is implemented
